remove stray `assert False` in order to fix #27
If I understand the comment in the while loop correctly, the purpose of
the while loop here is to throw away everything which is sent by the
sender to empty the UART buffer, and wait until a timeout occurs (`if
data is None`, i.e.  everything is sent, end of the current packet).
However, due to the `assert`, the loop terminates after the first byte,
unless it timeouts (which will most probably not be the case on current
connections).

Fixes #27