Read data as early and fast as possible
It makes a huge difference when reading from an embedded system if you
do things this way.

That changes things from:
Read a character
<twiddle thumbs>
Read the rest of the packet after possible buffer overrun has occurred
to:
Read everything upfront
<decide on what to do with that data>

This is good practice in general and, additionally, was required in
order to get this code to work in my case.